feat: add language profile page
