+++ src/pages/Languages/languages.styles.ts
@@ -23,12 +23,3 @@ export const Styledh2 = styled.h2`
   text-transform: capitalize;
   margin-bottom: 0.75rem;
 `;
-
-export const LanguagesContainer = styled.div`
-  margin-top: 2rem;
-  display: grid;
-  gap: 1rem;
-  padding: 1rem;
-  background-color: ${({ theme }) => theme.lightGrey};
-  border-radius: 0.5rem;
-`;
