+++ src/widgets/ui/users/LanguagesWidget/languagesWidget.styles.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const LanguagesContainer = styled.div`
+  margin-top: 2rem;
+  display: grid;
+  gap: 1rem;
+  padding: 1rem;
+  background-color: ${({ theme }) => theme.lightGrey};
+  border-radius: 0.5rem;
+`;
